reduce data usage, better data types

+++ data/left_contiguous_csr.py
@@ -1,5 +1,3 @@
-# Branch keshav
-
 import numpy as np
 
 
@@ -20,7 +18,7 @@ class LeftContiguousCSR:
     @staticmethod
     def load(filepath):
         npzfile = np.load(filepath)
-        return LeftContiguousCSR(npzfile['indptr'].astype(np.int64), npzfile['degrees'].astype(np.int64), npzfile['data'].astype(np.int64))
+        return LeftContiguousCSR(npzfile['indptr'], npzfile['degrees'], npzfile['data'])
 
 # uniform batch sampling can be done as follows
